pyproject.toml
[build-system]
requires = ["poetry-core=1.7.1"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "partition-registry"
description = "Library to work with Partition Registry"
authors = ["Vladislav Bakanov <rewerty52@ya.ru>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/vladislav-bakanov/partition-registry/blob/master/pyproject.toml"
version = "0.0.1"
classifiers = [
    "Development Status :: Beta",
    "Programming Language :: Python :: 3.11",
    "Intended Audience :: Developers :: Analytics",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
python = "3.11.7"
SQLAlchemy = "2.0.23"
# sqlalchemy-stubs = "0.4"
psycopg2-binary = "2.9.7"
fastapi = "0.104.1"
uvicorn = "0.24.0.post1"
pytz = ">=2023.3,<2023.4"
types-pytz = "2023.3.1.1"
httpx = "0.25.1"
python-dateutil = "2.8.2"
types-python-dateutil = "2.8.19.20240106"

[tool.poetry.dev-dependencies]
pytest = "8.0.0"
pytest-cov = "4.1.0"
hypothesis = "6.97.6"
mypy = "1.8.0"
pylint = "3.0.3"
